[TS] Change opcode types
